Refactor Code
* Move handlers out to a separate module
* Split `CalmApp._add_route` function
* General refactoring
