+++ calm/__init__.py
@@ -24,6 +24,6 @@ assigns URIs to them. Here is a basic example:
 For more information see `README.md`.
 """
 from calm.core import CalmApp as Application
+from calm.codec import ArgumentParser
 
-__all__ = ['Application']
 __version__ = '0.1.4'
